services/analyzer/requirements.txt
# Add service-specific dependencies here
openai
confluent-kafka
psycopg2-binary
PyYAML
python-dotenv